+++ samples/Axent.ExampleApi/OtherQuery.cs
@@ -1,4 +1,5 @@
 using Axent.Abstractions.Attributes;
+using Axent.Abstractions.Caching;
 using Axent.Abstractions.Models;
 using Axent.Abstractions.Options;
 using Axent.Abstractions.Pipelines;
@@ -13,6 +14,7 @@ internal sealed class OtherQuery : ICacheableQuery<OtherResponse>
     public required string Message { get; init; }
     public string CacheKey => $"{nameof(OtherQuery)}-{Message}";
     public bool BypassCache => false;
+    public CacheScope CacheScope => CacheScope.Culture;
     public CacheEntryOptions CacheOptions => new() { SlidingExpiration = TimeSpan.FromMinutes(5) };
 }